configure: Require OCaml and findlib if we need to run the generator.
When building from git, you need OCaml to compile the generator to
build the generated files.  Previously configure didn't detect that
situation, resulting in failed builds.  This commit causes the
configure script to error out if the generated files are not present
and OCaml is not installed.

Note that you can still build without OCaml, if you get the generated
files from somewhere else.  The source tarballs distributed on the
website contain the generated files.