launch: direct: Use a single -machine [type, ]accel=... option.
Previously we used (on arm): -M virt -machine accel=...

After this change, we use: -machine virt,accel=...
On x86 there is no change.

The -M option is an abbreviation for -machine, and we can combine the
two options together.  (Qemu combines the two options internally, but
it's more predictable if we do it ourselves).

Upstream qemu has supported the -machine option for over 5 years, and
in fact we implicitly relied on it (despite what the comment in the
code said).