v2v: move open_guestfs to Types
Move most of open_guestfs to Types, so a common function to open a
Guestfs handle is available.  Since it does not do all the things the
old open_guestfs did, the two operations have been moved to the only
place requiring them.

This function has been placed in Types and not Utils, as will be needed
by other functions in Types; Utils uses Types, so this avoids a circular
dependency.