launch: direct: aarch64: Add gic-version=host to -M option.
On ARM machines with vGICv3, we need to use the extra -M
parameter gic-version=3 (or gic-version=2 for the old vGIC).

qemu can detect this feature for us if we pass gic-version=host,
so add that to the -M option.  Unfortunately we cannot detect if
qemu itself supports this parameter, but this change only affects
aarch64.

Thanks: Wei Huang