v2v: Replace "=" by "_" in OVF .meta DESCRIPTION field.
The current VDSM .meta file parser is incorrect:

  https://code.engineering.redhat.com/gerrit/gitweb?p=vdsm.git;a=blob;f=vdsm/storage/fileVolume.py;h=20573bd2a3fe6909a8474bd7c53c8d0fea79091c;hb=c1289fd70edcde0a0b3bc8e37ecad1e2a5babe1a#l285

When parsing the key=value fields in the .meta file, it uses the
Python expression:

  key, value = l.split("=")

which throws an exception "ValueError: too many values to unpack" if
the value part also contains an "=" character (because split returns
an n-tuple with n > 2).  The correct way to split the string in this
situation is to use l.split("=", 1).

In commit 399e031c25a5385a3f918d0a604e2751bd34242f I changed the
DESCRIPTION field in .meta so that in RHEL it would look like this:

  DESCRIPTION=generated by libguestfs 1.32.2rhel=7,release=1.el7,libvirt

thus containing an "=" character in the value and throwing an
exception in VDSM.

Although this is a bug in VDSM which needs to be fixed separately,
change any occurrence of "=" in the value to "_" to avoid this
problem.

Thanks: Derrick Ornelas for helping to diagnose this bug and finding
the offending code in VDSM.