inspect: os-release: fix single-number version parsing (RHBZ#1321620)
Use the proper length modifier for asprintf, instead of a padding
modifier.  This fixes the parsing of versions in VERSION_ID which are a
single number (i.e. "X" and not "X.Y", etc).

Bug introduced with commit 32d19e3289bc259901f77398703f16cf6eabd510, but
uncovered recently with the switch away from VLAs, i.e.
commit 07c496c53c7835c0db882ef6761309db4952ebf6.