p2v: In non-GUI mode, make sure the progress of virt-p2v is visible (RHBZ#1229386).
By changing the logging mode of the p2v.service unit, we can make sure
that the progress of virt-p2v is visible in non-GUI mode at the
console.

Since the output is now always shown on the console, there's no need
to print the journalctl command for following that output when the
user logs in.