daemon: lvm: change the separator character to '\r'
Commit b91b39e06ab7eb9b9b06c8b4dfef2ef9f381ab19 changed the separator to
':', although this creates parsing issues when there are fields with
colons (for example lv_tags=imgbased:pool).

Change once more the separator, but this time using a non-printable
character such as '\r': while it will produce uglier debug logs, this
should greatly reduce the possibilities of conflicts with texts of
metadata.