v2v: Use separate scripts for the slow tests of real guests.
Following the same style used for the firstboot tests in the
customize/ subdirectory, use a separate test script for each real
guest that we test.  This has two advantages: (a) the tests run in
parallel (and are thus faster) and (b) automake "knows" about each
test and reports the results separately.