p2v: fix Gtk dependencies on Debian
On Debian, the names of the packages containing the Gtk 2 and 3
libraries have different patterns, and thus just the name will not work
for Gtk 3.

Apply a bit of m4 to use a different package name on Gtk 2 and other
versions.  It can be easily adapted to new models in future Gtk 4+,
should the naming pattern change from Gtk 3.