build: require ocaml-hivex for the daemon
Since commit d5b6f1df5ff2d387a5dfc89b8316c0dff67ce2c9, the daemon
requires ocaml-hivex; OTOH, nothing checks it is actually available, so
the build of the daemon will fail with a semi-cryptic error about the
lack of a directory in the OCaml install prefix.

As fix, check for the presence of the hivex module at build time,
failing earlier if not present.  The check is performed only when the
daemon is enabled, as ocaml-hivex is not used for anything else than the
daemon.