ruby: fix order of CFLAGS
Shuffle the order of the various CFLAGS used when building the Ruby
extension: first the flags from manywarnings (gnulib), then libguestfs
own CFLAGS, and then -DGUESTFS_PRIVATE=1 (so it is not overridden).

This matches also what happens in automake parts, and makes it possible
to override any flags from manywarnings also in the Ruby extension.