v2v: Add new Source_SATA for SATA disks.
Some hypervisors including KVM and VMware expose SATA as a separate
device emulation from IDE.  SATA devices are an evolution of IDE, but
in Linux they are handled like SCSI disks.

This change adds a new Source_SATA type to model this.  In the Linux
conversion module we treat them like SCSI (ie. with "sd" prefix).

It's not very clear how to handle them for target bus assignment, so I
arbitrarily mapped them to the SCSI bus.  It's possible we need a new
bus type to handle SATA properly.  As the code only applies to
removable devices it's not very important to get this right.