v2v: Fix parsing of OVA files and documentation for --network and --bridge (RHBZ#1559027).
When parsing OVA files we have to work backwards to whether the
original VMware object was a network or a bridge.  It seems as if the
presence of <rasd:Connection> is sufficient.  I dropped the use of
<rasd:ElementName> since it is literally just an internal name for the
network element, and is not significant in how the network is used.

Also update documentation because you must use ‘--bridge’ when
converting VMware guests to RHV.

For a full explanation see:

  https://bugzilla.redhat.com/show_bug.cgi?id=1559027#c2